refactor: move options menu to separated folder
